requirements.txt
gunicorn
blinker
Flask-SQLAlchemy
mysqlclient
Markdown
click